2_bbws_docs/project_plans/project-plan-4/stage-2-lambda-implementation/code-staging/requirements-dev.txt
# Marketing Lambda - Development Dependencies

# Include production dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
moto>=4.2.0  # AWS mocking

# Code quality
black>=23.12.0
ruff>=0.1.9
mypy>=1.8.0
boto3-stubs[dynamodb]>=1.34.0

# Type checking
types-boto3>=1.0.2
